a bit of cleanup
